Fixed all paths to work correctly on any platform
Using path.resolve in this way allows the right path to be generated on any platform... It figures out "/" and "\" on its own and creates the right path.

Other than that, there are just minor style fixes put in for consistency because eslint was complaining. I fixed the prod config file as well. I'll make a separate PR for that.